+++ pymkv/__init__.py
@@ -1,3 +1,4 @@
-from .pymkv import *
+from .MKVTrack import *
+from .MKVFile import *
 
 __version__ = '1.0.0'
